Working backwards compat script.
