+++ cmd.sh
@@ -7,8 +7,9 @@ if [ "$APP_ENV" == "DEVELOPMENT" ] || [ -z "$APP_ENV" ]; then
     trap "kill -9 $DATA_MODEL_MANAGER_PID" EXIT
 else
     MODE=$@
-    if [ "$MODE" == "--upgrade"]; then
-        python ./backwards_compatibility/104_to_110.py
+    if [ "$MODE" == "--upgrade" ]; then
+        python 104_to_110.py
+        exit 1
     fi
     if [ "$MODE" == "--data-model-manager" ]; then
         python data_model_manager_runner.py
